[ECHOJS-250] [ECHOJS-251] Done
